% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_annotation.R
\name{assemble_annotation}
\alias{assemble_annotation}
\title{Assemble a single EML annotation element}
\usage{
assemble_annotation(annotation_row)
}
\arguments{
\item{annotation_row}{(data.frame) Data frame with one row containing info for the annotation}
}
\value{
(list) emld list for annotation
}
\description{
Assemble a single EML annotation element
}
