% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_boilerplate.R
\name{assemble_boilerplate}
\alias{assemble_boilerplate}
\title{Assemble boilerplate items.}
\usage{
assemble_boilerplate(bp_df, bp_people, bp_setting)
}
\arguments{
\item{bp_df}{(data.frame) Data frame containing boilerplate XML trees on access, dataset-level distribution, and project.}

\item{bp_people}{(data.frame) Data frame containing expanded info on contact, metadata provider, publisher.}

\item{bp_setting}{(character) The boilerplate setting ("default" or otherwise).}
}
\value{
(list) Named list containing boilerplate items: scope, system, access, project, distribution, contact, metadata provider, publisher, license.
}
\description{
Assemble boilerplate items.
}
