% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_coverage.R
\name{assemble_coverage}
\alias{assemble_coverage}
\title{Assemble dataset coverage elements}
\usage{
assemble_coverage(meta_list)
}
\arguments{
\item{meta_list}{(list) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}
}
\value{
(list) A named list containing geographicCoverage, temporalCoverage, and taxonomicCoverage elements, each NULL if no information is provided.
}
\description{
Assemble geographic, temporal, and taxonomic coverage elements at the dataset level.
}
