% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_keywords.R
\name{assemble_keyword}
\alias{assemble_keyword}
\title{Assemble a single keyword.}
\usage{
assemble_keyword(keyword)
}
\arguments{
\item{keyword}{(data.frame) A data.frame with one row.}
}
\value{
(list) An emld list for a single keyword with keywordType if present.
}
\description{
Assemble a single keyword.
}
