% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_keywords.R
\name{assemble_keywordset}
\alias{assemble_keywordset}
\title{Assemble keywordSets by thesaurus.}
\usage{
assemble_keywordset(keyword_df)
}
\arguments{
\item{keyword_df}{(data.frame) A data.frame containing keywords. Required columns: "keyword", "keyword_thesaurus", "keywordtype".}
}
\value{
(list) An emld list structure containing as many keywordSets as there are thesauri in input.
}
\description{
Assemble keywordSets by thesaurus.
}
