% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_maintenance.R
\name{assemble_maintenance}
\alias{assemble_maintenance}
\title{Assemble dataset level maintenance tree}
\usage{
assemble_maintenance(dataset_df, changehistory_df)
}
\arguments{
\item{dataset_df}{(data.frame) A data frame containing information on the dataset.}

\item{changehistory_df}{(data.frame) A data frame containing information on the dataset's change history.}
}
\value{
(list) An emld list structure containting the EML maintenance tree: description, update frequency, change history
}
\description{
Assemble dataset level maintenance tree
}
