% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_person.R
\name{assemble_person}
\alias{assemble_person}
\title{Assemble EML list structure for singular personnel.}
\usage{
assemble_person(person)
}
\arguments{
\item{person}{(data.frame) A single-row data.frame containing information on a single ResponsibleParty.}
}
\value{
(list) emld list structure.
}
\description{
Assemble emld list structure for an EML ResponsibleParty type.
}
