% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_person.R
\name{assemble_personnel}
\alias{assemble_personnel}
\title{Create a person list structure}
\usage{
assemble_personnel(personnel_df)
}
\arguments{
\item{personnel_df}{(data.frame) A data frame containing information on personnel}
}
\description{
Create a person list structure
}
