% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_publications.R
\name{assemble_publications}
\alias{assemble_publications}
\title{Assemble EML list structure for multiple publications}
\usage{
assemble_publications(publication_df)
}
\arguments{
\item{publication_df}{(data.frame) A data frame containing information on publications}
}
\value{
(list) emld list with unnamed child items, each with information on publications
}
\description{
Assemble emld list structure for multiple citationType.
}
