% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_taxonomy.R
\name{assemble_taxon}
\alias{assemble_taxon}
\title{Assemble a single non-recursive, non-nested taxonomicClassification node}
\usage{
assemble_taxon(taxa_row)
}
\arguments{
\item{taxa_row}{(data.frame) One row of a data.frame containing information on a single taxon.}
}
\value{
(list) An emld formatted unnamed list describing a single taxon
}
\description{
Non-recursive or non-nested means this function does not lookup the full taxonomic tree and nest the leaf taxon within.
}
