% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_taxonomy.R
\name{assemble_taxon_nested}
\alias{assemble_taxon_nested}
\title{Assemble a single recursive nested taxonomicClassification node}
\usage{
assemble_taxon_nested(classification, providerurl, commonname = NULL)
}
\arguments{
\item{classification}{(data.frame) One data.frame containing one full taxonomic classification tree for one leaf taxon.}

\item{provider}{(character) URL of the taxonomic authority/provider}
}
\value{
(list) A nested list structure that represents the taxonomic classification tree
}
\description{
Recursive or nested means this function does use the full taxonomic tree and nest the leaf taxon within.
}
