% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_keywords.R
\name{assemble_thesaurus}
\alias{assemble_thesaurus}
\title{Assemble a single thesaurus's keywordSet}
\usage{
assemble_thesaurus(thesaurus)
}
\arguments{
\item{thesaurus}{(data.frame) A data.frame with keywords in one thesaurus.}
}
\value{
(list) An emld list for one keywordSet corresponding to the thesaurus.
}
\description{
Assemble a single thesaurus's keywordSet
}
