% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_person.R
\name{assemble_userid}
\alias{assemble_userid}
\title{Assemble userId EML element}
\usage{
assemble_userid(person)
}
\arguments{
\item{person}{(data.frame) One row of a data.frame containing information on a single ResponsibleParty corresponding to a metabase name ID.}
}
\value{
(list) Emld list structure
}
\description{
Assemble userId EML element
}
