% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EML.R
\name{create_EML}
\alias{create_EML}
\title{Create EML list object.}
\usage{
create_EML(meta_list, entity_list, dataset_id, boilerplate_path,
  license_path)
}
\arguments{
\item{meta_list}{A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{entity_list}{(character) A list of entities returned by \code{\link{create_entity_all}}.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{boilerplate_path}{(character) System path to XML file containing boilerplate items.}

\item{license_path}{(character) System path to pandoc compatible file containing intellectual rights statement.}
}
\value{
(list) A list containing all EML elements. Supply this list object to \code{\link[EML]{eml_validate}} and \code{\link[EML]{write_eml}} to validate and write to .xml file.
}
\description{
Create ready-to-validate-and-write EML list object.
}
\examples{
\dontrun{
# continued from \\code{\\link{get_meta}} and \\code{\\link{create_entity_all}}
EML <- create_EML(meta_list = metadata, entity_list = entities, dataset_id = 1, boilerplate_path = here::here("documents", "boilerplate.xml"), license_path = here::here("documents", "license.docx"))
}
}
