% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entity_all.R
\name{create_entity_all}
\alias{create_entity_all}
\title{Quickly create all EML entity list objects.}
\usage{
create_entity_all(meta_list, file_dir = getwd(), dataset_id,
  entity_numbers = NULL, skip_checks = FALSE)
}
\arguments{
\item{meta_list}{(character) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{file_dir}{(character) Path to directory containing flat files (data files). Defaults to current R working directory.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{entity_numbers}{(numeric) Vector of entity numbers to include. Defaults to all of them. Use only when you need to delibrately exclude a couple entities.}

\item{skip_checks}{(logical) Whether to skip checking for attribute congruence. Defaults to FALSE.}
}
\value{
(list) A list containing all data entities from dataset. Use this in the `entity_list` argument for \code{\link{create_EML}}.
First level list elements are grouped by entity types present in dataset and named accordingly.
Each first level element is a list of unnamed lists; the number of elements correspond to how many entities of each type are present in dataset.
Second level elements are analogous to output from \code{\link{create_entity}}.
}
\description{
. Use to quickly create EML entity list objects from all entities listed in dataset.
}
\examples{
\dontrun{
# continued from \\code{\\link{get_meta}}
entities <- create_entity_all(meta_list = metadata, dataset_id = 1)
}

}
