% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methods.R
\name{create_method_step}
\alias{create_method_step}
\title{Create method step.}
\usage{
create_method_step(step_id, meta_list, dataset_id, file_dir = "")
}
\arguments{
\item{step_id}{(numeric) A methodStep ID. Distinct IDs constitute distinct methodSteps.}

\item{meta_list}{(list) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{file_dir}{(character) Path to directory containing flat files (abstract and method documents). Defaults to "" or current R working directory.}
}
\value{
(list) An emld list containing a methodStep for the specified dataset ID and methodStep ID.
}
\description{
Create an emld list for a methodStep.
}
