% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{get_meta}
\usage{
get_meta(dbname, schema = "mb2eml_r", dataset_ids, host = "localhost",
  port = 5432, user = NULL, password = NULL)
}
\arguments{
\item{dbname}{(character) name of database.}

\item{schema}{(character) name of schema containing views, defaults to 'mb2eml_r'.}

\item{dataset_ids}{(numeric) Number or numeric vector of dataset ids to query.}

\item{host}{(numeric) host name, defaults to 'localhost'.}

\item{port}{(character) port number, defaults to 5432.}

\item{user}{(character) username, optional. If NULL, RStudio will create a pop-up asking for username.}

\item{password}{(character) password to user, optional. If NULL, RStudio will create a pop-up asking for username.}
}
\value{
(list) A list of data frames corresponding to views from specified schema in metabase
to pass to [create_entity()] and [create_EML()]
}
\description{
Function to connect to Postgres metabase and query for metadata
}
\examples{
Query multiple dataset at once
metadata <- get_meta(dbname = "ble_metabase", dataset_ids = c(1, 2))

}
