% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_maintenance.R
\name{make_history}
\alias{make_history}
\title{Create a single entry in dataset change history.}
\usage{
make_history(changehistory)
}
\arguments{
\item{changehistory}{(data.frame) A data.frame with a single row.}
}
\value{
(list) An emld list for one change history entry.
}
\description{
Create a single entry in dataset change history.
}
