% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_taxonomy.R
\name{match_provider}
\alias{match_provider}
\title{Resolve provider ID or URL based on the information provided in metabase}
\usage{
match_provider(taxa_row, type = "url")
}
\arguments{
\item{taxa_row}{(data.frame) One row of a data.frame containing information on a single taxon.}

\item{type}{(character) "url" or "id". Type of information to return. id returns an abbreviation that can be used to plug in many taxonomic functions. Defaults to url.}
}
\value{
(character) URL of the taxonomic provider as can be best guessed by the function. "unknown" if no provider url or name provided (nothing to go on), or if the provider name is not known.
}
\description{
Resolve provider ID or URL based on the information provided in metabase
}
