% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_coverage.R
\name{assemble_coverage}
\alias{assemble_coverage}
\title{Assemble dataset coverage elements}
\usage{
assemble_coverage(meta_list, expand_taxa = FALSE, skip_taxa = FALSE)
}
\arguments{
\item{meta_list}{(list) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{expand_taxa}{(logical) TRUE/FALSE on whether assemble_taxonomic will lookup and fully expand a leaf node taxon's full taxonomic classification (kingdom to the lowest rank provided) into nested EML taxonomicCoverage elements (TRUE) or simply make a taxonomic coverage module based on the information provided in metabase (FALSE). This assumes, of course, that the taxa provided are only the leaf nodes. If so, setting this to TRUE and having the full classification may help your dataset be more discover-able, however the lookup process may be more prone to errors. If this is set to TRUE, rows containing taxa from unsupported providers, or from supported providers but whose classification lookups fail, will not be expanded. The function will use information from the taxonid, taxonrankvalue, taxonid_provider, and (if you have it) providerurl columns from the vw_eml_taxonomy view queried from metabase. It expects taxonid to contain the correct identifier for the taxon from the listed taxonomic authority/provider, taxonrankvalue to contain the taxon's name, taxonid_provider to provide a correctly spelled name or commonly used ID for the taxonomic provider/authority (e.g. ITIS for the Integrated Taxonomy Information System), and providerurl to contain a working url to the same.  Defaults to FALSE}

\item{skip_taxa}{(logical) Whether to skip the call to \code{assemble_taxonomic}. Provided in case assemble_taxonomic fails in some way -- taxonomies are tricky; one option is to manually insert in a text editor a snippet of EML generated elsewhere, into the complete EML output from MetaEgress. Defaults to FALSE.}
}
\value{
(list) A named list containing geographicCoverage, temporalCoverage, and taxonomicCoverage elements, each NULL if no information is provided.
}
\description{
Assemble geographic, temporal, and taxonomic coverage elements at the dataset level.
}
