% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_person.R
\name{assemble_person}
\alias{assemble_person}
\title{Assemble EML list structure for singular personnel.}
\usage{
assemble_person(nameid)
}
\arguments{
\item{nameid}{(data.frame) A data.frame containing information on a single ResponsibleParty corresponding to a metabase name ID. Most often single row but can contain multiple rows if there are multiple user IDs listed.}
}
\value{
(list) emld list structure.
}
\description{
Assemble emld list structure for an EML ResponsibleParty type.
}
