% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_person.R
\name{assemble_personnel}
\alias{assemble_personnel}
\title{Assemble EML list structure for multiple personnel.}
\usage{
assemble_personnel(personnel_df)
}
\arguments{
\item{personnel_df}{(data.frame) A data frame containing information on personnel}
}
\value{
(list) emld list with unnamed child items, each with information on personnel
}
\description{
Assemble emld list structure for multiple EML ResponsibleParty type.
}
