% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EML.R
\name{create_EML}
\alias{create_EML}
\title{Create EML.}
\usage{
create_EML(
  meta_list,
  entity_list,
  dataset_id,
  file_dir = getwd(),
  expand_taxa = FALSE,
  skip_taxa = FALSE,
  ble_options = FALSE
)
}
\arguments{
\item{meta_list}{(list) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{entity_list}{(character) A list of entities returned by \code{\link{create_entity_all}}.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{file_dir}{(character) Path to directory containing flat files (abstract and method documents). Defaults current R working directory.}

\item{expand_taxa}{(logical) TRUE/FALSE on whether assemble_taxonomic will lookup and fully expand a leaf node taxon's full taxonomic classification (kingdom to the lowest rank provided) into nested EML taxonomicCoverage elements (TRUE) or simply make a taxonomic coverage module based on the information provided in metabase (FALSE). This assumes, of course, that the taxa provided are only the leaf nodes. If so, setting this to TRUE and having the full classification may help your dataset be more discover-able, however the lookup process may be more prone to errors. If this is set to TRUE, rows containing taxa from unsupported providers, or from supported providers but whose classification lookups fail, will not be expanded. The function will use information from the taxonid, taxonrankvalue, taxonid_provider, and (if you have it) providerurl columns from the vw_eml_taxonomy view queried from metabase. It expects taxonid to contain the correct identifier for the taxon from the listed taxonomic authority/provider, taxonrankvalue to contain the taxon's name, taxonid_provider to provide a correctly spelled name or commonly used ID for the taxonomic provider/authority (e.g. ITIS for the Integrated Taxonomy Information System), and providerurl to contain a working url to the same.}

\item{skip_taxa}{(logical) Whether to skip the call to \code{assemble_taxonomic}. Provided in case assemble_taxonomic fails in some way -- taxonomies are tricky; one option is to manually insert in a text editor a snippet of EML generated elsewhere, into the complete EML output from MetaEgress. Defaults to FALSE.}

\item{ble_options}{(logical) Whether to perform tasks specific to BLE-LTER: add an additional metadata snippet to facilitate replication to the Arctic Data Center. Defaults to FALSE.}
}
\value{
(list) An EML package-compatible XML list tree. Supply this list object to \code{\link[EML]{eml_validate}} and \code{\link[EML]{write_eml}} to, in order, validate and write to .xml file.
}
\description{
Create an EML package-compatible XML list tree for an EML document, ready to validate and write to .xml file.
}
\examples{
\dontrun{
# continued from \code{\link{get_meta}} and \code{\link{create_entity_all}}
EML <- create_EML(meta_list = metadata, entity_list = entities, dataset_id = 1, license_path = here::here("documents", "license.docx"))
}
}
