% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EML.R
\name{create_EML}
\alias{create_EML}
\title{create_eml}
\usage{
create_EML(meta_list, dataset_id, boilerplate_path, license_path,
  data_table, other_entity = NULL)
}
\arguments{
\item{meta_list}{A list of dataframes containing metadata returned by [get_meta()].}

\item{dataset_id}{A number for dataset ID.}

\item{boilerplate_path}{A system path to XML file containing boilerplate items.}

\item{license_path}{A system path to pandoc compatible file containing intellectual rights statement.}

\item{data_table}{A list of one or many dataTable objects returned by [create_entity()].}

\item{other_entity}{A list of one or many otherEntity objects returned by [create_entity()].}
}
\description{
Function to create ready-to-validate-and-write EML list object
}
\examples{
\dontrun{continued from [get_meta()] and [create_entity()]
metadata <- get_meta(dbname = "ble_metabase", dataset_ids = c(1, 2))
entity_1 <- create_entity(meta_list = metadata, dataset_id = 1, entity = 1)

use lapply to loop through many entities. Separate data tables from other entities.
dt <- c(1:4)
other <- c(5:7)
data_tables <- lapply(dt, create_entity, meta_list = metadata, dataset_id = 1)
other_entities <- lapply(other, create_entity, meta_list = metadata, dataset_id = 1)
}

}
