% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entity.R
\name{create_entity}
\alias{create_entity}
\title{Create EML entity list object.}
\usage{
create_entity(
  meta_list,
  file_dir = getwd(),
  dataset_id,
  entity,
  skip_checks = FALSE
)
}
\arguments{
\item{meta_list}{(character) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{file_dir}{(character) Path to directory containing flat files (data files). Defaults to current R working directory. Note: if there's information on "entityrecords", "filesize", "filesize_units", and "checksum" columns in the entities table in metabase, there is no need for reading the actual files, so this field can stay NULL.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{entity}{(numeric) An entity number.}

\item{skip_checks}{(logical) Whether to skip checking for attribute congruence. Defaults to FALSE.}
}
\value{
(list) A list object containing one data entity.
}
\description{
Use to examine entity list structure or troubleshoot invalid EML, or to put together custom entity lists. Use \code{\link{create_entity_all}} for common EML generation usage, which loops this function over all entities listed in a dataset under the hood.
}
