% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entity.R
\name{create_entity}
\alias{create_entity}
\title{Create EML entity list object.}
\usage{
create_entity(meta_list, file_dir = NULL, dataset_id, entity)
}
\arguments{
\item{meta_list}{(character) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{file_dir}{(character) Path to directory containing flat files (data files). Defaults to current R working directory if NULL. Note: if the "entityrecords", "filesize", "filesize_units", and "checksum" columns in the entities table in metabase are filled out, there is no need for reading the actual files, so this field can stay NULL.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{entity}{(numeric) An entity number.}
}
\value{
(list) A list object containing one data entity.
}
\description{
Use to examine entity list structure or troubleshoot invalid EML, or to put together custom entity lists. Use \code{\link{create_entity_all}} for common EML generation usage, which calls this function under the hood.
}
\examples{
\dontrun{
# continued from \\code{\\link{get_meta}}
# A single entity. Useful to examine EML list structure and troubleshoot.
entity_1 <- create_entity(meta_list = metadata, dataset_id = 1, entity = 1)

# Many entities. Loop separately for each entity type and name accordingly.
data_tables <- c(1:4)
other_entities <- c(5:7)
entity_list <- list(
  dataTable = lapply(data_tables, create_entity, meta_list = metadata, dataset_id = 1),
  otherEntity = lapply(other_entities, create_entity, meta_list = metadata, dataset_id = 1)
)
}
}
