% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methods.R
\name{create_method_section}
\alias{create_method_section}
\title{Create dataset method section.}
\usage{
create_method_section(meta_list, dataset_id, file_dir = "")
}
\arguments{
\item{meta_list}{(list) A list of dataframes containing metadata returned by \code{\link{get_meta}}.}

\item{dataset_id}{(numeric) A dataset ID.}

\item{file_dir}{(character) Path to directory containing flat files (abstract and method documents). Defaults to current R working directory if "".}
}
\value{
(list) An emld list containing information on methods in LTER-core-metabase for the specified dataset ID.
}
\description{
Create emld list for a dataset method section.
}
